refactor(#3805): split EmptyValueType into types/empty_value.rs (campsite rule)
types.rs 2920 -> 3004 (was 3220): the tag enum, its CqlType mapping and the
source-derived membership rule move to their own module beside
types/comparator.rs (epic #1116). The Value::Empty variant and its
ordering/rendering arms stay with the rest of Value.
